add default_network_id to token respone

+++ app/controllers/devise/oauth2_providable/tokens_controller.rb
@@ -6,8 +6,15 @@ class Devise::Oauth2Providable::TokensController < ApplicationController
     if oauth2_current_refresh_token then
       @refresh_token = oauth2_current_refresh_token
     else
-      Devise::Oauth2Providable::RefreshToken.where(:account_id=>current_account.id,:client_id=>[1,2]).delete_all
-      @refresh_token = oauth2_current_client.refresh_tokens.create!(:account_id => current_account.id)
+
+      if oauth2_current_client then
+        Devise::Oauth2Providable::RefreshToken.where(:account_id=>current_account.id,:client_id=>[1,2]).delete_all
+        @refresh_token = oauth2_current_client.refresh_tokens.create!(:account_id => current_account.id)
+      else
+        return render(:json => {:errors=>{:message=>"Oauth client not found.",:status_code=>:invalid_request}},:status => 400)
+
+      end
+
     end
 
     # Hard code for delete android and ios,1 for ios,2 for android
@@ -28,7 +35,10 @@ class Devise::Oauth2Providable::TokensController < ApplicationController
     Rails.cache.delete "/oauth2/access_token_by_account/#{current_account.id}"
     old_tokens.each {|x| Rails.cache.delete "/oauth2/access_token/#{x}" }
 
-    render :json => @access_token.token_response
+    token_resp = @access_token.token_response
+    token_resp.merge!(:default_network_id => current_account.home_user.network_id)
+    
+    render :json => token_resp
   end
 
   private
@@ -36,6 +46,7 @@ class Devise::Oauth2Providable::TokensController < ApplicationController
   def oauth2_current_client
     env[Devise::Oauth2Providable::CLIENT_ENV_REF]
   end
+
   def oauth2_current_refresh_token
     env[Devise::Oauth2Providable::REFRESH_TOKEN_ENV_REF]
   end
